ostree: patch to setup a curl request timeout
Set a timeout for an overall request processing in order to avoid an
ostree pull/fetch API call or CLI command hang.  The timeout value can
be specified via an env variable OSTREE_CURL_TIMEOUT, if not defined, or
equal 0 or it cannot be converted to a long type then the default value
is set.

Based on 2b742b96341a245b2a1c66bb93a837345e533c63 in meta-lmp; see
https://github.com/foundriesio/meta-lmp/pull/382 and
https://github.com/foundriesio/meta-lmp/pull/391.

Authored-by: Mike Sul <mike.sul@foundries.io>
Signed-off-by: Patti Vacek <pattivacek@gmail.com>