classes/sota: no need to add wic when sota is enabled
While most boards will indeed use wic format for images, it is not
something that is strictly required when using sota, as we do have
some BSPs that require only the ext4 image instead (e.g. tegra).

Wic should be included, when needed, at the machine config instead (or
via sota_<machine>.bbclass).

To avoid breaking compatibility with the current machines, wic is also
added to machine-specific classes that requires it.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>