nfs-utils: prefer tmpfiles instead of a custom service
The directories that are critical for nfs-utils at runtime are
/var/lib/nfs, /var/lib/nfs/v4recovery and /var/lib/nfs/statd, which
can easily be created during boot via a systemd tmpfiles, avoiding
the need for a custom service.

While this can be applied in a generic way, having it applied only when
sota is enabled as /var is not preserved only when used with sota/ostree.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>