image_types_ostree: adds ostree_rmdir_helper
This introduces an helper to delete folders on the rootfs.

The helper can be configured with the variable OSTREE_RMDIR_HELPER_MSGTYPE
to generate a bitbake bbwarn or bbfatal message when the directory is not empty.
Also changed all log messages to the default type to be consistent.
The default value is bbwarn that make it backward compatible keeping the removal
of the files/directories as before.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>