fit-conf: fix u-boot boot fail
On board using a fitImage, poky's upstream change from the following
commit:
https://git.yoctoproject.org/poky/commit/?id=ab6b5e97cebe19938baa403da63
07ca320294b3a
inserted a regression of the fit-conf.bb recipe.
Causing a U-boot boot fault displaying this type of message:
```
  Could not find configuration node
  ERROR: can't get kernel image!
  Unknown command '!' - try 'help'
  resetting ...
```
fitImage configuration labels are now prefixed with conf- instead of
conf@.
This commit incorporates this modification for the fit-conf.bb recipe.