ostree-kernel-initramfs: deltask instaed of using noexec flag
When the 'noexec' flag was used, the falged task is not executed
however the the tasks flaged in 'depends' are.

The consequence of this is that the kernel and its modules was
compiled whenever this recipe needs to run the do_intall task.
This beacuse the module-base.bbclass triggers the make-mod-scripts
task do_compile.

| $ grep depends meta/classes-recipe/module-base.bbclass
| do_configure[depends] += "make-mod-scripts:do_compile"

Deleting the task also cleanup theirs dependencies so do_configure
no longer pulls its dependencies and consequently the kernel is no
longer rebuild.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>