ostree-kernel-initramfs: pick the kernel-abiversion from deploy dir
The KERNEL_VERSION was populated from with the content of
the file kernel-abiversion available in the STAGING_KERNEL_BUILDDIR.
It turns out that this file is not stored on sstate-cache and
is available only after the running of the do_shared_workdir task
from kernel.bbclass, which implies that the kernel has to be
compiled.

This change now uses what exists in the task kernel:do_deploy that is
available in sstate-cache.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>
(cherry picked from commit 40b5f0dba4a2c2201da26faffefa2be0e1d09870)