image_types_ota: introduce generic repository configuration
This allows one to set the OTA ostree repository (i.e. the one belonging
to a sysroot) configuration. The configuration is defined by the new
variable OSTREE_OTA_REPO_CONFIG which would be set to a string in the
form:

"key1:value1 key2:value2 ..."

The settings are applied to the repository via "ostree config set".

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>