image_types_ostree: refactor garagesign task
To enable usage of the Uptane's 'ota-tuf' repository for uptane-sign, we
need to change this task. Since in more recent releases there, the tool
name has changed from 'garage-sign' to 'uptane-sign'.

This way, we set the name of the tool used in a customizable variable,
defaulting to 'garage-sign', and if anyone wants to use the latest
version of this tool (given Aktualizr was also refactored), they just
need to set this new variable with the tool name.

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 397c3ab0ab41a4918aec1ba67474940bfb1cb923)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>