classes: ostree: remove content from /boot
Content from /boot ends up getting commited into the initial ostree
revision, but it is not really used after deployment, since ostree is
the one managing the /boot content. Remove content from such folder as
it is not really used by the system, just extra space which is not used
(including kernel and bootloader files which are not from ostree).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>