dbus: let systemd-tmpfiles handle creation of statedir
OE-core recipe [1] explicitly creates /var/lib/dbus and assigns
user and group for it to messagebus:messagebus. dbus already has
a tmpfiles [2] that creates the machine-id symlink and /var/lib/dbus
directory if not present.
Create another tmpfiles that creates this directory at runtime and
assigns the user group as per OE-core.
Remove the directory if it is empty at build time to
fix warnings when using systemd.

[1] https://gitlab.freedesktop.org/dbus/dbus/-/blob/dbus-1.16/bus/tmpfiles.d/dbus.conf.in?ref_type=heads#L9
[2] https://git.openembedded.org/openembedded-core/tree/meta/recipes-core/dbus/dbus_1.16.2.bb?h=walnascar#n162

Signed-off-by: Vishwas Udupa <vudupa@qti.qualcomm.com>