systemd/sd-boot: backport patch to fix support for empty initrd
Empty initrds are used both with fitimages and ukis (as it is packaged
together with the kernel), but recent systemd crashes when loading empty
initrd files. Backport upstream systemd patch which fixes the handling
of empty initrd files.

Patch applied on both systemd and systemd-boot as both recipes are
currently generating systemd-boot (one in EFI and another in /usr/lib),
which needs fixing in oe-core.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>